% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_generate.R
\name{check_missing_inputs}
\alias{check_missing_inputs}
\title{Identify which referenced files are missing from available files}
\usage{
check_missing_inputs(referenced_files, available_files)
}
\arguments{
\item{referenced_files}{Character vector of basenames referenced by generated code.}

\item{available_files}{Character vector of basenames available locally.}
}
\value{
Character vector of missing basenames.
}
\description{
Identify which referenced files are missing from available files
}
\keyword{internal}
