% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_openai.R
\name{delete_file_from_openai}
\alias{delete_file_from_openai}
\title{Delete a file from the OpenAI Files API}
\usage{
delete_file_from_openai(
  file_id,
  api_key = Sys.getenv("OPENAI_API_KEY"),
  api_base = "https://api.openai.com/v1"
)
}
\arguments{
\item{file_id}{Character scalar. File id to delete.}

\item{api_key}{Character scalar. OpenAI API key.}

\item{api_base}{Character scalar. API base URL.}
}
\value{
TRUE invisibly on success.
}
\description{
Delete a file from the OpenAI Files API
}
\keyword{internal}
