% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_scanner.R
\name{list_code_files}
\alias{list_code_files}
\title{List code/text files under a directory}
\usage{
list_code_files(
  dir,
  recursive = TRUE,
  code_ext = c(".R", ".r", ".Rmd", ".qmd", ".py", ".do", ".sas", ".sps", ".txt", ".md")
)
}
\arguments{
\item{dir}{Source directory.}

\item{recursive}{Logical. Recurse into subdirectories?}

\item{code_ext}{Character vector of file extensions (including dot).}
}
\value{
Character vector of file paths.
}
\description{
List code/text files under a directory
}
\keyword{internal}
