% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_scanner.R
\name{list_data_files}
\alias{list_data_files}
\title{List data files under a directory}
\usage{
list_data_files(
  dir,
  recursive = TRUE,
  data_ext = c(".csv", ".tsv", ".txt", ".xlsx", ".xls", ".sav", ".dta", ".rds", ".RData")
)
}
\arguments{
\item{dir}{Source directory.}

\item{recursive}{Logical. Recurse into subdirectories?}

\item{data_ext}{Character vector of file extensions (including dot).}
}
\value{
Character vector of file paths.
}
\description{
List data files under a directory
}
\keyword{internal}
