% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_openai.R
\name{list_openai_files}
\alias{list_openai_files}
\title{List files in the OpenAI Files API}
\usage{
list_openai_files(
  api_key = Sys.getenv("OPENAI_API_KEY"),
  api_base = "https://api.openai.com/v1"
)
}
\arguments{
\item{api_key}{Character scalar. OpenAI API key.}

\item{api_base}{Character scalar. API base URL.}
}
\value{
A data frame with at least \code{id}, \code{filename}, \code{purpose}.
}
\description{
List files in the OpenAI Files API
}
\keyword{internal}
